test/requirements.txt
selenium==3.141
msedge-selenium-tools
